% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggonion.R
\name{ggonion}
\alias{ggonion}
\title{Draw onion diagram}
\usage{
ggonion(x, ratio = 2, bias = 0, color = NA)
}
\arguments{
\item{x}{vector of labels for onion diagramm}

\item{ratio}{(optional) rmodificator of nested circle size, default is 2}

\item{bias}{(optional) vertical bias for nested circles, default is 0}

\item{color}{(optional) vector of colors for circles}
}
\description{
Draw onion diagram
}
