% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggonion.R
\name{iterative_get_y0}
\alias{iterative_get_y0}
\title{Hlpr to iterative calculate y coord bias for circles center using previous circle precalculated value}
\usage{
iterative_get_y0(df, bias)
}
\arguments{
\item{bias}{bias [0:1]}
}
\description{
Hlpr to iterative calculate y coord bias for circles center using previous circle precalculated value
}
