% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggonion.R
\name{iterative_get_y_label}
\alias{iterative_get_y_label}
\title{Calculate labels y coord as middle between upper half of circles}
\usage{
iterative_get_y_label(df)
}
\value{
df with y_label coord
}
\description{
Calculate labels y coord as middle between upper half of circles
}
