% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggonion.R
\name{make_color}
\alias{make_color}
\title{Check color vector if provided, if not - generate colors}
\usage{
make_color(prep_circles_df, color)
}
\arguments{
\item{prep_circles_df}{df from prepare_circles_df function}

\item{color}{color vector}
}
\value{
color vector
}
\description{
Check color vector if provided, if not - generate colors
}
