% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggonion.R
\name{prepare_circles_df}
\alias{prepare_circles_df}
\title{Prepare dataframe with circles parameters}
\usage{
prepare_circles_df(x, ratio, bias)
}
\arguments{
\item{x}{vector of classes for onion diagramm}

\item{ratio}{modificator of nested circle size}

\item{bias}{vertical bias of nested circles}
}
\value{
df[label,x0,y0,r,x_label,y_label]
}
\description{
Prepare dataframe with circles parameters
}
